Added install screen with 2 buttons (explore and from git)
