+++ .idea/modules.xml
@@ -2,8 +2,6 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/Netbeast.iml" filepath="$PROJECT_DIR$/Netbeast.iml" />
-      <module fileurl="file://$PROJECT_DIR$/NetbeastApp.iml" filepath="$PROJECT_DIR$/NetbeastApp.iml" />
       <module fileurl="file://$PROJECT_DIR$/android.iml" filepath="$PROJECT_DIR$/android.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
